+++ sct_asar_reader/src/sct_asar_reader/protocol_implementation.py
@@ -189,12 +189,6 @@ class ASARChannelManager:
             # should be a 1D array
             self._lines_per_burst_array = np.repeat(self._channel.raster_info.lines, 1)
 
-        # pulse rate
-        if self._channel.pulse is not None:
-            self._signal_pulse_rate = self._channel.pulse.bandwidth / self._channel.pulse.pulse_length
-        else:
-            self._signal_pulse_rate = 1
-
         # prf
         self._prf = self._channel.swath_info.prf
 
@@ -357,11 +351,6 @@ class ASARChannelManager:
     def sampling_constants(self) -> SARSamplingFrequencies:
         """Channel data signal sampling frequencies"""
         return self._channel.sampling_constants
-
-    @property
-    def pulse_rate(self) -> float:
-        """Signal pulse rate"""
-        return self._signal_pulse_rate
 
     @property
     def looking_side(self) -> SARSideLooking: