+++ sct_biomass_reader/src/sct_biomass_reader/__init__.py
@@ -10,4 +10,4 @@ from sct_biomass_reader.interface import get_ale_corrector, get_detector, get_ma
 
 __all__ = ["get_ale_corrector", "get_detector", "get_manager"]
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"