+++ sct_eos04_reader/src/sct_eos04_reader/protocol_implementation.py
@@ -195,9 +195,6 @@ class EOS04ChannelManager:
             # should be a 1D array
             self._lines_per_burst_array = np.repeat(self._channel.raster_info.lines.length, 1)
 
-        # pulse rate
-        self._signal_pulse_rate = self._channel.pulse.bandwidth / self._channel.pulse.pulse_length
-
         # prf
         self._prf = self._channel.swath_info.prf
 
@@ -363,11 +360,6 @@ class EOS04ChannelManager:
     def sampling_constants(self) -> SARSamplingFrequencies:
         """Channel data signal sampling frequencies"""
         return self._channel.sampling_constants
-
-    @property
-    def pulse_rate(self) -> float:
-        """Signal pulse rate"""
-        return self._signal_pulse_rate
 
     @property
     def looking_side(self) -> SARSideLooking: