+++ sct_saocom_reader/src/sct_saocom_reader/__init__.py
@@ -10,4 +10,4 @@ from sct_saocom_reader.interface import get_ale_corrector, get_detector, get_man
 
 __all__ = ["get_manager", "get_detector", "get_ale_corrector"]
 
-__version__ = "1.1.1"
+__version__ = "1.1.2"