+++ sct_saocom_reader/src/sct_saocom_reader/protocol_implementation.py
@@ -192,12 +192,6 @@ class SAOCOMChannelManager:
             # should be a 1D array
             self._lines_per_burst_array = np.repeat(self._channel.raster_info.lines.length, 1)
 
-        # pulse rate
-        if self._channel.pulse is not None:
-            self._signal_pulse_rate = self._channel.pulse.bandwidth_hz / self._channel.pulse.length_s
-        else:
-            self._signal_pulse_rate = 1
-
         # prf
         self._prf = self._channel.swath_info.prf
 
@@ -363,11 +357,6 @@ class SAOCOMChannelManager:
     def sampling_constants(self) -> SARSamplingFrequencies:
         """Channel data signal sampling frequencies"""
         return self._channel.sampling_constants
-
-    @property
-    def pulse_rate(self) -> float:
-        """Signal pulse rate"""
-        return self._signal_pulse_rate
 
     @property
     def looking_side(self) -> SARSideLooking: